feat: separate 'activate'
